OD20DEVAT3-69, OD20DEVAT3-29, OD20DEVAT3-53. Changes according to front-end.

+++ src/main/resources/frontend/angular-app/src/app/entity/Directory.ts
@@ -1,5 +1,7 @@
+import {BigInteger} from "@angular/compiler/src/i18n/big_integer";
+
 export class Directory {
-  objectId: bigint;
+  objectId: BigInteger;
   path: string;
   enabled: boolean;
   lastExistenceCheck: Date;
